websocket_client/requirements.txt
Flask
Flask-SocketIO
eventlet
pika
SQLAlchemy
psycopg2-binary
Flask-SQLAlchemy
python-dotenv
requests
flask-socketio